feat(license): add reusable license details card component
