+++ app/screens/AppInfoScreens/LicenseScreen.tsx
@@ -1,29 +1,12 @@
 import React from 'react';
-import { ScrollView, Text, View, Linking, Alert } from 'react-native';
-import { StyledHeader } from '@components';
-import { UnistylesRuntime, createStyleSheet, useStyles } from 'react-native-unistyles';
+import { ScrollView } from 'react-native';
+import { StyledHeader, StyledLicenseDetailsCard } from '@components';
+import { createStyleSheet, useStyles } from 'react-native-unistyles';
+import { UnistylesRuntime } from 'react-native-unistyles';
 import licenses from '../../assets/oss-licenses.json';
 
 const LicensesScreen = () => {
   const { styles } = useStyles(stylesheet);
-
-  const handleLinkPress = (url: string) => {
-    Alert.alert(
-      'Open External Link',
-      'You are about to open an external website. Do you want to continue?',
-      [
-        {
-          text: 'Cancel',
-          style: 'cancel'
-        },
-        {
-          text: 'Open',
-          onPress: () => Linking.openURL(url)
-        }
-      ],
-      { cancelable: true }
-    );
-  };
 
   return (
     <>
@@ -43,23 +26,14 @@ const LicensesScreen = () => {
           const version = packageName.slice(lastAtIndex + 1);
 
           return (
-            <View key={index} style={styles.licenseMainContainer}>
-              <Text style={styles.libraryNameText}>{name}</Text>
-              <Text style={styles.versionText}>Version: {version}</Text>
-              <Text style={styles.licenseText}>License: {details.licenses}</Text>
-              <Text
-                style={styles.repositoryText}
-                onPress={() => handleLinkPress(details.repository)}
-              >
-                Repository: {details.repository}
-              </Text>
-              <Text
-                style={styles.licenseUrlText}
-                onPress={() => handleLinkPress(details.licenseUrl)}
-              >
-                License URL: {details.licenseUrl}
-              </Text>
-            </View>
+            <StyledLicenseDetailsCard
+              key={index}
+              name={name}
+              version={version}
+              license={details.licenses}
+              repository={details.repository}
+              licenseUrl={details.licenseUrl}
+            />
           );
         })}
       </ScrollView>
@@ -67,7 +41,7 @@ const LicensesScreen = () => {
   );
 };
 
-const stylesheet = createStyleSheet(({ colors, fonts }) => ({
+const stylesheet = createStyleSheet(({ colors }) => ({
   mainContainer: {
     flex: 1,
     backgroundColor: colors.backgroundColor,
@@ -75,45 +49,6 @@ const stylesheet = createStyleSheet(({ colors, fonts }) => ({
   },
   scrollContentContainer: {
     paddingBottom: UnistylesRuntime.insets.bottom * 2,
-  },
-  licenseMainContainer: {
-    width: '100%',
-    backgroundColor: colors.card,
-    paddingVertical: (UnistylesRuntime.screen.height * 2) / 100,
-    paddingHorizontal: (UnistylesRuntime.screen.width * 5) / 100,
-    borderRadius: (UnistylesRuntime.screen.height * 1) / 100,
-    marginBottom: (UnistylesRuntime.screen.height * 2) / 100,
-  },
-  libraryNameText: {
-    color: colors.accent,
-    fontSize: 14,
-    fontFamily: fonts.Nunito_Black,
-    marginBottom: (UnistylesRuntime.screen.height * 0.5) / 100,
-  },
-  versionText: {
-    fontSize: 10,
-    color: colors.card_typography,
-    fontFamily: fonts.Montserrat_Medium,
-    marginBottom: (UnistylesRuntime.screen.height * 0.5) / 100,
-  },
-  licenseText: {
-    fontSize: 10,
-    color: colors.card_typography,
-    fontFamily: fonts.Montserrat_Medium,
-    marginBottom: (UnistylesRuntime.screen.height * 0.5) / 100,
-  },
-  repositoryText: {
-    fontSize: 10,
-    color: colors.accent,
-    fontFamily: fonts.Montserrat_Medium,
-    textDecorationLine: 'underline',
-    marginBottom: (UnistylesRuntime.screen.height * 0.5) / 100,
-  },
-  licenseUrlText: {
-    fontSize: 10,
-    color: colors.accent,
-    fontFamily: fonts.Montserrat_Medium,
-    textDecorationLine: 'underline',
   },
 }));
